Pull from rodent data from online repo
Given that our computers have different paths to the data, I changed the data source to the online repo. But I left your line in if there's something special about your version of the data. We can also braoden this function so that we call tell it where to read from, if we need that.