Portalr friendly rodent abundance processing
This commit finishes the process of taking the original hand-crafted rodent data extraction and processing and makes a process built on the portalr package. Still to do: remove get data from the get_rodent so that data extraction doesn't have to be repeated for each type of data summary one wishes.